Create SceneViewerQuad.py
Four scene.Scenes embedded in one ui.View.  Touch the four quadrants of the screen to create colored circles.  I want to drag a circle across multiple scenes and have it change colors based on the underlying scene.  Unfortunately, touch events are only triggered in the scene that they start in.